requirements.txt
absl-py==0.9.0
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pathspec==0.8.0
pylint==2.5.3
pymongo==3.11.0
PyYAML==5.3.1
regex==2020.7.14
registry==0.4.2
requests==2.24.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
